requirements.txt
accelerate==0.28.0
fastapi==0.116.1
numpy==1.26.0
opencv-python-headless==4.11.0.86
pdf2image==1.17.0
pillow==9.4.0
pymongo==4.13.2
python-multipart==0.0.20
qwen-vl-utils==0.0.11
torch==2.1.0
transformers==4.53.1
uvicorn==0.35.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0